backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain>=0.0.267
langchain-community>=0.0.1
langchain-google-genai>=0.0.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openai>=1.0.0
